Php-fpm socket is working.
I added a test case with the host machine.
The files must be in the same container as the socket so the fpm can process the requests.
Whenever adding we do https://localhost:443/index.html a pagina sera reencaminhada para http://localhost:8080/index.html.

This is the standart behaviour of the browser.